requirements.txt
click==8.1.7
pyfiglet==1.0.2
rich==13.7.1
websocket_client==1.7.0
